requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==23.12.1
certifi==2023.11.17
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
lxml==5.1.0
mypy-extensions==1.0.0
openai==1.8.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
prompt-toolkit==1.0.14
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyInquirer==1.0.3
python-docx==1.1.0
python-dotenv==1.0.0
regex==2023.12.25
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
wcwidth==0.2.13
